requirements.txt
fastmcp
web3
python-dotenv
httpx